Simplify `PlatformCertFile::get()`.
